Form merge

+++ src/Illuminate/Foundation/Console/DownCommand.php
@@ -34,6 +34,12 @@ class DownCommand extends Command
     public function handle()
     {
         try {
+            if (file_exists(storage_path('framework/down'))) {
+                $this->comment('Application is already down.');
+
+                return true;
+            }
+
             file_put_contents(storage_path('framework/down'),
                               json_encode($this->getDownFilePayload(),
                               JSON_PRETTY_PRINT));
